Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/parse-error-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'application/json' is not a valid JavaScript MIME type.

Harness Error (TIMEOUT), message = null

TIMEOUT JSON modules: parse error Test timed out
